refactor: rename tests
Move tests and resources out of postgresql/mysql packages and rename
them to have PostgreSQL/MySQL prefix instead. This makes it clearer
which class broke when there are test failures.
