+++ v2/sourcedb-to-spanner/src/test/java/com/google/cloud/teleport/v2/templates/MySQLForeignKeyDependencyIT.java
@@ -16,16 +16,15 @@
 package com.google.cloud.teleport.v2.templates;
 
 import static org.apache.beam.it.truthmatchers.PipelineAsserts.assertThatResult;
+import static org.junit.Assert.assertEquals;
 
 import com.google.cloud.teleport.metadata.SkipDirectRunnerTest;
 import com.google.cloud.teleport.metadata.TemplateIntegrationTest;
-import java.util.List;
-import java.util.Map;
+import java.util.Arrays;
 import org.apache.beam.it.common.PipelineLauncher;
 import org.apache.beam.it.common.PipelineOperator;
 import org.apache.beam.it.common.utils.ResourceManagerUtils;
 import org.apache.beam.it.gcp.spanner.SpannerResourceManager;
-import org.apache.beam.it.gcp.spanner.matchers.SpannerAsserts;
 import org.apache.beam.it.jdbc.MySQLResourceManager;
 import org.junit.After;
 import org.junit.Before;
@@ -33,37 +32,26 @@ import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
- * An integration test for {@link SourceDbToSpanner} Flex template which tests a single sharded
- * migration on a simple schema.
+ * An integration test for {@link SourceDbToSpanner} Flex template which tests FK dependency
+ * migration.
  */
 @Category({TemplateIntegrationTest.class, SkipDirectRunnerTest.class})
 @TemplateIntegrationTest(SourceDbToSpanner.class)
 @RunWith(JUnit4.class)
-public class SingleShardIT extends SourceDbToSpannerITBase {
+public class MySQLForeignKeyDependencyIT extends SourceDbToSpannerITBase {
+  private static final Logger LOG = LoggerFactory.getLogger(MySQLForeignKeyDependencyIT.class);
   private static PipelineLauncher.LaunchInfo jobInfo;
 
   public static MySQLResourceManager mySQLResourceManager;
   public static SpannerResourceManager spannerResourceManager;
 
-  private static final String MYSQL_DUMP_FILE_RESOURCE =
-      "SingleShardWithTransformation/mysql-schema.sql";
+  private static final String MYSQL_DUMP_FILE_RESOURCE = "ForeignKeyDependencyIT/mysql-schema.sql";
 
-  private static final String SPANNER_DDL_RESOURCE =
-      "SingleShardWithTransformation/spanner-schema.sql";
-
-  private static final String SESSION_FILE_RESOURCE = "SingleShardWithTransformation/session.json";
-
-  private static final String TABLE = "SingleShardWithTransformationTable";
-
-  private static final String PKID = "pkid";
-
-  private static final String NAME = "name";
-
-  private static final String STATUS = "status";
-
-  private static final String SHARD_ID = "migration_shard_id";
+  private static final String SPANNER_DDL_RESOURCE = "ForeignKeyDependencyIT/spanner-schema.sql";
 
   /**
    * Setup resource managers and Launch dataflow job once during the execution of this test class. \
@@ -80,18 +68,14 @@ public class SingleShardIT extends SourceDbToSpannerITBase {
     ResourceManagerUtils.cleanResources(spannerResourceManager, mySQLResourceManager);
   }
 
-  /**
-   * TODO: This IT is currently not complete since shard id population is pending on reader. This
-   * test needs to be updated whenever reader support is added.
-   */
   @Test
-  public void singleShardWithIdPopulationTest() throws Exception {
+  public void linearDependencyTest() throws Exception {
     loadSQLFileResource(mySQLResourceManager, MYSQL_DUMP_FILE_RESOURCE);
     createSpannerDDL(spannerResourceManager, SPANNER_DDL_RESOURCE);
     jobInfo =
         launchDataflowJob(
             getClass().getSimpleName(),
-            SESSION_FILE_RESOURCE,
+            null,
             null,
             mySQLResourceManager,
             spannerResourceManager,
@@ -100,17 +84,12 @@ public class SingleShardIT extends SourceDbToSpannerITBase {
     PipelineOperator.Result result = pipelineOperator().waitUntilDone(createConfig(jobInfo));
     assertThatResult(result).isLaunchFinished();
 
-    SpannerAsserts.assertThatStructs(
-            spannerResourceManager.readTableRecords(TABLE, PKID, NAME, STATUS, SHARD_ID))
-        .hasRecordsUnorderedCaseInsensitiveColumns(getExpectedData());
-  }
-
-  private List<Map<String, Object>> getExpectedData() {
-    return List.of(
-        Map.of(PKID, 1, NAME, "Alice", STATUS, "active", SHARD_ID, "NULL"),
-        Map.of(PKID, 2, NAME, "Bob", STATUS, "inactive", SHARD_ID, "NULL"),
-        Map.of(PKID, 3, NAME, "Carol", STATUS, "pending", SHARD_ID, "NULL"),
-        Map.of(PKID, 4, NAME, "David", STATUS, "complete", SHARD_ID, "NULL"),
-        Map.of(PKID, 5, NAME, "Emily", STATUS, "error", SHARD_ID, "NULL"));
+    for (String tableName :
+        Arrays.asList("t1", "t2", "t3", "t4", "t5", "t6", "t7", "t8", "t9", "t10")) {
+      assertEquals(
+          "Asserting count for " + tableName,
+          10L,
+          spannerResourceManager.getRowCount(tableName).longValue());
+    }
   }
 }
