+++ v2/sourcedb-to-spanner/src/test/java/com/google/cloud/teleport/v2/templates/MySQLSingleShardIT.java
@@ -15,11 +15,10 @@
  */
 package com.google.cloud.teleport.v2.templates;
 
-import com.google.cloud.spanner.Struct;
+import static org.apache.beam.it.truthmatchers.PipelineAsserts.assertThatResult;
+
 import com.google.cloud.teleport.metadata.SkipDirectRunnerTest;
 import com.google.cloud.teleport.metadata.TemplateIntegrationTest;
-import com.google.common.collect.ImmutableList;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import org.apache.beam.it.common.PipelineLauncher;
@@ -34,27 +33,37 @@ import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
- * An integration test for {@link SourceDbToSpanner} Flex template which tests a basic migration on
- * a simple schema.
+ * An integration test for {@link SourceDbToSpanner} Flex template which tests a single sharded
+ * migration on a simple schema.
  */
 @Category({TemplateIntegrationTest.class, SkipDirectRunnerTest.class})
 @TemplateIntegrationTest(SourceDbToSpanner.class)
 @RunWith(JUnit4.class)
-public class SessionSchemaMapperIT extends SourceDbToSpannerITBase {
-  private static final Logger LOG = LoggerFactory.getLogger(SessionSchemaMapperIT.class);
-  private static final HashSet<SessionSchemaMapperIT> testInstances = new HashSet<>();
+public class MySQLSingleShardIT extends SourceDbToSpannerITBase {
   private static PipelineLauncher.LaunchInfo jobInfo;
 
   public static MySQLResourceManager mySQLResourceManager;
   public static SpannerResourceManager spannerResourceManager;
 
-  private static final String SESSION_FILE_RESOURCE = "SchemaMapperIT/company-session.json";
-  private static final String MYSQL_DDL_RESOURCE = "SchemaMapperIT/company-mysql-schema.sql";
-  private static final String SPANNER_DDL_RESOURCE = "SchemaMapperIT/company-spanner-schema.sql";
+  private static final String MYSQL_DUMP_FILE_RESOURCE =
+      "SingleShardWithTransformation/mysql-schema.sql";
+
+  private static final String SPANNER_DDL_RESOURCE =
+      "SingleShardWithTransformation/spanner-schema.sql";
+
+  private static final String SESSION_FILE_RESOURCE = "SingleShardWithTransformation/session.json";
+
+  private static final String TABLE = "SingleShardWithTransformationTable";
+
+  private static final String PKID = "pkid";
+
+  private static final String NAME = "name";
+
+  private static final String STATUS = "status";
+
+  private static final String SHARD_ID = "migration_shard_id";
 
   /**
    * Setup resource managers and Launch dataflow job once during the execution of this test class. \
@@ -71,43 +80,37 @@ public class SessionSchemaMapperIT extends SourceDbToSpannerITBase {
     ResourceManagerUtils.cleanResources(spannerResourceManager, mySQLResourceManager);
   }
 
+  /**
+   * TODO: This IT is currently not complete since shard id population is pending on reader. This
+   * test needs to be updated whenever reader support is added.
+   */
   @Test
-  public void noTransformationTest() throws Exception {
-    loadSQLFileResource(mySQLResourceManager, MYSQL_DDL_RESOURCE);
+  public void singleShardWithIdPopulationTest() throws Exception {
+    loadSQLFileResource(mySQLResourceManager, MYSQL_DUMP_FILE_RESOURCE);
     createSpannerDDL(spannerResourceManager, SPANNER_DDL_RESOURCE);
     jobInfo =
         launchDataflowJob(
             getClass().getSimpleName(),
             SESSION_FILE_RESOURCE,
-            "mapper",
+            null,
             mySQLResourceManager,
             spannerResourceManager,
             null,
             null);
     PipelineOperator.Result result = pipelineOperator().waitUntilDone(createConfig(jobInfo));
+    assertThatResult(result).isLaunchFinished();
 
-    List<Map<String, Object>> companyMySQL =
-        mySQLResourceManager.runSQLQuery("SELECT company_id, company_name FROM company");
-    ImmutableList<Struct> companySpanner =
-        spannerResourceManager.readTableRecords("company", "company_id", "company_name");
+    SpannerAsserts.assertThatStructs(
+            spannerResourceManager.readTableRecords(TABLE, PKID, NAME, STATUS, SHARD_ID))
+        .hasRecordsUnorderedCaseInsensitiveColumns(getExpectedData());
+  }
 
-    SpannerAsserts.assertThatStructs(companySpanner)
-        .hasRecordsUnorderedCaseInsensitiveColumns(companyMySQL);
-
-    List<Map<String, Object>> employeeMySQL =
-        mySQLResourceManager.runSQLQuery(
-            "SELECT employee_id, company_id, employee_name, employee_address FROM employee");
-    ImmutableList<Struct> employeeSpanner =
-        spannerResourceManager.readTableRecords(
-            "employee", "employee_id", "company_id", "employee_name", "employee_address");
-
-    SpannerAsserts.assertThatStructs(employeeSpanner)
-        .hasRecordsUnorderedCaseInsensitiveColumns(employeeMySQL);
-
-    ImmutableList<Struct> employeeAttribute =
-        spannerResourceManager.readTableRecords(
-            "employee_attribute", "employee_id", "attribute_name", "value");
-
-    SpannerAsserts.assertThatStructs(employeeAttribute).hasRows(4); // Supports composite keys
+  private List<Map<String, Object>> getExpectedData() {
+    return List.of(
+        Map.of(PKID, 1, NAME, "Alice", STATUS, "active", SHARD_ID, "NULL"),
+        Map.of(PKID, 2, NAME, "Bob", STATUS, "inactive", SHARD_ID, "NULL"),
+        Map.of(PKID, 3, NAME, "Carol", STATUS, "pending", SHARD_ID, "NULL"),
+        Map.of(PKID, 4, NAME, "David", STATUS, "complete", SHARD_ID, "NULL"),
+        Map.of(PKID, 5, NAME, "Emily", STATUS, "error", SHARD_ID, "NULL"));
   }
 }
