+++ v2/sourcedb-to-spanner/src/test/resources/DataTypesIT/mysql-data-types.sql
