+++ v2/sourcedb-to-spanner/src/test/resources/DataTypesIT/mysql-spanner-schema.sql
