+++ v2/sourcedb-to-spanner/src/test/resources/DataTypesIT/postgresql-spanner-schema.sql
