+++ v2/sourcedb-to-spanner/src/test/resources/SourceDbToSpannerLT/postgresql-spanner-schema.sql
